+++ EShop.Presentation/ProductDto.cs
@@ -1,0 +1,4 @@
+﻿namespace EShop.Presentation
+{
+    public record ProductDto(string Name, decimal Price);
+}
